feat(footer): dsfr footer
