+++ package.json
@@ -12,10 +12,12 @@
   "dependencies": {
     "@babel/core": "^7.7.4",
     "@babel/runtime": "^7.6.2",
+    "@dataesr/react-dsfr": "^1.1.11",
     "@datapunt/matomo-tracker-react": "0.5.1",
     "@date-io/date-fns": "1.x",
     "@emotion/react": "^11.8.1",
     "@emotion/styled": "^11.8.1",
+    "@gouvfr/dsfr": "^1.4.1",
     "@mui/icons-material": "^5.4.2",
     "@mui/lab": "^5.0.0-alpha.70",
     "@mui/material": "^5.4.3",
