+++ web/root.js
@@ -7,10 +7,6 @@ import {
   useLocation,
   useHistory
 } from "react-router-dom";
-
-import "./index.css";
-import "common/assets/styles/root.scss";
-
 import {
   StoreSyncedWithLocalStorageProvider,
   useStoreSyncedWithLocalStorage
@@ -49,6 +45,9 @@ import {
 import CircularProgress from "@mui/material/CircularProgress";
 import { ErrorBoundary } from "./common/ErrorFallback";
 import { RegulationDrawerContextProvider } from "./landing/ResourcePage/RegulationDrawer";
+
+import "@gouvfr/dsfr/dist/dsfr/dsfr.min.css";
+import "common/assets/styles/root.scss"; // should be imported after dsfr
 
 const matomo = createInstance({
   urlBase: "https://stats.data.gouv.fr",
